query brightness on initialization
The following adjusts the iManager's backlight control module to query
the existing brightness level of a device on initialization (instead of
assuming/forcing the initial brightness to maximum).

The original implementation would force a device's brightness to maximum
(during the probe stage) and assume maximum brightness during the
initialization phase. By removing the forced "brightness to maximum"
during the probe stage, a system can continue to use the already
configured brightness of a device from a previously loaded backlight
module event or an external/independent configuration event. With this
change, the iManager driver cannot assume the initial device's
brightness is at maximum; thus, the additional change during
initialization which queries the brightness level on start-up.

Signed-off-by: James Knight <james.knight@rockwellcollins.com>